+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next version
 
+- Add JSON schema for assets
+- Add support for RouterProxyFactory contract
+- Add deployments of MangroveOrder, MangroveOrder-Router, and RouterProxyFactory v2.0.0-b1.0
+
 # 2.0.0
 
 - Add dependencies to deployments. This allows querying for e.g. periphery or strat contracts that are tied to a specific Mangrove instance.